Initial Project Setup (#18)
This is the first drop of the project that lays the groundwork to actually build out the necessary work.

CMake generator for Windows/VS and GGP/Ninja
CMake + Python to generate GGP/VS solution/project
Scripts to deploy and build configs
Simplified starter app that has a spinning cube
vkex used as Vulkan interface
glfw for windowing, even though glfw stinks

~~~~~~~~~~~~~~~~
Commit messages that got squashed!

* Add expected vkex dependencies

These are the known submodule dependencies!

* Initial vkex drop

Grabbed from https://github.com/chaoticbob/VulkanTestSuite @ 1c9582b628066fd504d6b32e19fa2a87a6f4c855

* Strip out GPA references in vkex code

Just do a dumb teardown

* Grab assets

* source shaders

* First pass at CMake setup

* Get vkex and app building

Had to manually plumb Vulkan headers + lib paths
Also had to fix up usage of an old IMGUI Vulkan API. Might need to check the submodule VulkanTestSuite is at, and maybe roll back if there bugs

* Fix IMGUI + VK interface

IMGUI needs an image count, so we grab the swapchain count

* Update gitignore + vscode settings

* Borrow Application support code from VulkanTestSuite

This is the first drop that builds in Windows 64 with VS 2017!!

* Remove Windows version reporting

It was using a deprecated API. Not sure why we need a specific Windows version. If we care about it, we should re-factor the code

* Add GGP GLFW mini-drop

Right now we have a custom GLFW version that's only present in the SDK samples. Until we have it as part of the actual SDK or the open-source GLFW project, it's not very easy to include the GGP GLFW anywhere.

So instead, we just jam in the pre-build release static lib, plus the headers.

* Clean up CMake files

* Add build info to README

* Update required CMake version

* Add pre-built libglfw to tracking

* Prepare makefiles for GGP VS generation

+ strip out tinyobjloader

* Initial GGP VS post-process script

README updated with instructions, but I need to generate bash scripts to do this work!

* Deploy + build scripts