VK_EXT_sample_locations implementation + documentation (#69)
Added support for VK_EXT_sample_locations, which simplifies the logic of the checkerboard implementation, though using the extension adds in a fair amount of complexity that isn't very appealing.

Also, there's a doc detailing how the extension works, and how it was utilized in the sample. Spoiler alert: I kinda hate this extension.