Make AMD_shader_core_properties optional
Previously, if you were on an AMD GPU, vkex _required_ AMD_shader_core_properties. If the extension happens to be missing, the app will no longer error out.

There is another point (in PhysicalDevice creation) where the props are checked. In that code, it uses CreateDevice, but can absorb the failure fine.